Relocate block arrays from quadrants into players
All entities in a corner now live on the player model.
If we want, this can allow us to keep colors consistent.
This reduces collision checks (mostly regarding blocks).
This makes for easier defeated statis changes accross paddles and player 'zones'
(as well as anything else we may want to implement upon defeat or victory)